Use the maximum between elapsed time and 1
Most likely the elapsed time will always be >= 1,
so taking the minimum between that and 1 will always result in 1.
But we just don't want to end up with 0 (or a negative value), so we
should take the maximum.

Changed the values in the test because for both 1 and 2ms the correction
will be 16, but for higher values it will change.